+++ MangoBot-cli/src/main/java/io/github/mangomaner/mangobot/model/dto/message/QueryMessagesByMessageIdRequest.java
@@ -10,11 +10,11 @@ public class QueryMessagesByMessageIdRequest {
 
     @Schema(description = "Bot ID")
     @NotNull(message = "Bot ID不能为空")
-    private Integer botId;
+    private Long botId;
 
     @Schema(description = "群组ID或好友ID")
     @NotNull(message = "群组ID或好友ID不能为空")
-    private Integer targetId;
+    private Long targetId;
 
     @Schema(description = "消息ID")
     @NotNull(message = "消息ID不能为空")
