+++ MangoBot-cli/src/main/java/io/github/mangomaner/mangobot/model/dto/message/SearchMessagesRequest.java
@@ -11,11 +11,11 @@ public class SearchMessagesRequest {
 
     @Schema(description = "Bot ID")
     @NotNull(message = "Bot ID不能为空")
-    private Integer botId;
+    private Long botId;
 
     @Schema(description = "群组ID或好友ID")
     @NotNull(message = "群组ID或好友ID不能为空")
-    private Integer targetId;
+    private Long targetId;
 
     @Schema(description = "搜索关键词")
     @NotBlank(message = "搜索关键词不能为空")
